unit fixes in the bw limiter

+++ dcpp/DownloadManager.cpp
@@ -356,7 +356,7 @@ void DownloadManager::endData(UserConnection* aSource) {
 	checkDownloads(aSource);
 }
 
-size_t DownloadManager::throttleGetSlice()  {
+int64_t DownloadManager::throttleGetSlice()  {
 	if (mThrottleEnable) {
 		int64_t left = mDownloadLimit - getRunningAverage();
 		if (-left >= mDownloadLimit)  {
@@ -367,11 +367,11 @@ size_t DownloadManager::throttleGetSlice()  {
 			return mByteSlice;
 		}
 	} else {
-		return (size_t)-1;
+		return -1;
 	}
 }
 
-size_t DownloadManager::throttleCycleTime() {
+uint32_t DownloadManager::throttleCycleTime() {
 	if (mThrottleEnable)
 		return mCycleTime;
 	return 0;
